layers: Remove core_validation.cpp
