+++ BUILD.gn
@@ -116,7 +116,6 @@ core_validation_sources = [
   "layers/core_checks/cmd_buffer_dynamic_validation.cpp",
   "layers/core_checks/cmd_buffer_validation.cpp",
   "layers/core_checks/copy_blit_resolve_validation.cpp",
-  "layers/core_checks/core_validation.cpp",
   "layers/core_checks/core_validation.h",
   "layers/core_checks/descriptor_validation.cpp",
   "layers/core_checks/device_memory_validation.cpp",
