+++ scripts/vk_validation_stats.py
@@ -45,7 +45,6 @@ ignore_unassigned = True # These are not found in layer code unless they appear 
 layer_source_files = [common_codegen.repo_relative(path) for path in [
     'layers/core_checks/android_validation.cpp',
     'layers/core_checks/buffer_validation.cpp',
-    'layers/core_checks/core_validation.cpp',
     'layers/state_tracker/cmd_buffer_state.cpp', # some Video VUIDs are in here
     'layers/core_checks/cmd_buffer_dynamic_validation.cpp',
     'layers/core_checks/cmd_buffer_validation.cpp',
